app/services/llm_service.py
import asyncio
import json
from typing import Dict, Any, Optional
import logging
from app.core.config import settings
from app.models.schemas import LLMProvider, CVEvaluation, ProjectEvaluation
from app.core.exceptions import LLMServiceError
from app.core.retry import retry_async, LLM_RETRY_CONFIG

# Import LLM clients
try:
    import openai
    OPENAI_AVAILABLE = True
except ImportError:
    OPENAI_AVAILABLE = False

try:
    import google.generativeai as genai
    GEMINI_AVAILABLE = True
except ImportError:
    GEMINI_AVAILABLE = False

class LLMService:
    def __init__(self):
        self.setup_clients()
    
    def setup_clients(self):
        """Setup LLM clients based on available API keys"""
        if OPENAI_AVAILABLE and settings.OPENAI_API_KEY:
            openai.api_key = settings.OPENAI_API_KEY
        
        if GEMINI_AVAILABLE and settings.GEMINI_API_KEY:
            genai.configure(api_key=settings.GEMINI_API_KEY)
    
    async def evaluate_cv(self, cv_text: str, job_requirements: str, context: str = "") -> CVEvaluation:
        """Evaluate CV with fallback to mock data on failure"""
        try:
            prompt = self._create_cv_prompt(cv_text, job_requirements, context)
            
            if self.provider == LLMProvider.OPENAI:
                response = await self._call_openai(prompt)
            elif self.provider == LLMProvider.GEMINI:
                response = await self._call_gemini(prompt)
            else:  # MOCK
                return self._get_mock_cv_evaluation()
            
            parsed_response = self._parse_json_response(response, "cv")
            return CVEvaluation(**parsed_response)
            
        except (LLMServiceError, Exception) as e:
            logging.warning(f"LLM evaluation failed, using mock data: {str(e)}")
            return self._get_mock_cv_evaluation()

    async def evaluate_project(self, project_text: str, job_requirements: str, context: str = "") -> ProjectEvaluation:
        """Evaluate project with fallback to mock data on failure"""
        try:
            prompt = self._create_project_prompt(project_text, job_requirements, context)
            
            if self.provider == LLMProvider.OPENAI:
                response = await self._call_openai(prompt)
            elif self.provider == LLMProvider.GEMINI:
                response = await self._call_gemini(prompt)
            else:  # MOCK
                return self._get_mock_project_evaluation()
            
            parsed_response = self._parse_json_response(response, "project")
            return ProjectEvaluation(**parsed_response)
            
        except (LLMServiceError, Exception) as e:
            logging.warning(f"LLM evaluation failed, using mock data: {str(e)}")
            return self._get_mock_project_evaluation()

    async def generate_summary(self, cv_eval: CVEvaluation, project_eval: ProjectEvaluation, job_requirements: str) -> str:
        """Generate overall summary with fallback"""
        try:
            prompt = self._create_summary_prompt(cv_eval, project_eval, job_requirements)
            
            if self.provider == LLMProvider.OPENAI:
                response = await self._call_openai(prompt)
            elif self.provider == LLMProvider.GEMINI:
                response = await self._call_gemini(prompt)
            else:  # MOCK
                return self._get_mock_summary()
            
            return response.strip()
            
        except (LLMServiceError, Exception) as e:
            logging.warning(f"Summary generation failed, using mock summary: {str(e)}")
            return self._get_mock_summary()
    
    def _create_cv_evaluation_prompt(self, cv_text: str, job_description: str) -> str:
        """Create prompt for CV evaluation"""
        return f"""
        Evaluate the following CV against the job requirements and return a JSON response with the following structure:
        {{
            "technical_skills_match": 0.0-1.0,
            "experience_level": 0.0-1.0,
            "relevant_achievements": 0.0-1.0,
            "cultural_fit": 0.0-1.0,
            "cv_match_rate": 0.0-1.0,
            "cv_feedback": "detailed feedback string"
        }}
        
        Job Description:
        {job_description}
        
        CV Content:
        {cv_text}
        
        Evaluate based on:
        1. Technical Skills Match (backend, databases, APIs, cloud, AI/LLM exposure)
        2. Experience Level (years, project complexity)
        3. Relevant Achievements (impact, scale)
        4. Cultural Fit (communication, learning attitude)
        
        Return only valid JSON.
        """
    
    def _create_project_evaluation_prompt(self, project_text: str, study_case_brief: str) -> str:
        """Create prompt for project evaluation"""
        return f"""
        Evaluate the following project report against the study case requirements and return a JSON response with the following structure:
        {{
            "correctness": 1-5,
            "code_quality": 1-5,
            "resilience": 1-5,
            "documentation": 1-5,
            "creativity_bonus": 1-5,
            "project_score": 1-5,
            "project_feedback": "detailed feedback string"
        }}
        
        Study Case Brief:
        {study_case_brief}
        
        Project Report:
        {project_text}
        
        Evaluate based on:
        1. Correctness (meets requirements: prompt design, chaining, RAG, handling errors)
        2. Code Quality (clean, modular, testable)
        3. Resilience (handles failures, retries)
        4. Documentation (clear README, explanation of trade-offs)
        5. Creativity/Bonus (optional improvements like authentication, deployment, dashboards)
        
        Return only valid JSON.
        """
    
    def _create_summary_prompt(self, cv_result: Dict, project_result: Dict) -> str:
        """Create prompt for overall summary"""
        return f"""
        Based on the CV evaluation and project evaluation results below, provide a concise overall summary of the candidate's fit for the backend developer position.
        
        CV Results: {json.dumps(cv_result)}
        Project Results: {json.dumps(project_result)}
        
        Provide a 2-3 sentence summary highlighting strengths and areas for improvement.
        """
    
    async def _evaluate_with_openai(self, prompt: str, eval_type: str) -> Dict[str, Any]:
        """Evaluate using OpenAI"""
        response = await self._call_openai(prompt)
        return self._parse_json_response(response, eval_type)
    
    async def _evaluate_with_gemini(self, prompt: str, eval_type: str) -> Dict[str, Any]:
        """Evaluate using Gemini"""
        response = await self._call_gemini(prompt)
        return self._parse_json_response(response, eval_type)
    
    @retry_async(LLM_RETRY_CONFIG)
    async def _call_openai(self, prompt: str, model: str = None) -> str:
        """Call OpenAI API with retry logic"""
        try:
            model = model or settings.OPENAI_MODEL
            response = await openai.ChatCompletion.acreate(
                model=model,
                messages=[{"role": "user", "content": prompt}],
                max_tokens=2000,
                temperature=0.3
            )
            return response.choices[0].message.content.strip()
        except Exception as e:
            logging.error(f"OpenAI API call failed: {str(e)}")
            raise LLMServiceError(f"OpenAI API call failed: {str(e)}", {"model": model})

    @retry_async(LLM_RETRY_CONFIG)
    async def _call_gemini(self, prompt: str, model: str = None) -> str:
        """Call Gemini API with retry logic"""
        try:
            model_name = model or settings.GEMINI_MODEL
            model = genai.GenerativeModel(model_name)
            response = await model.generate_content_async(
                prompt,
                generation_config=genai.types.GenerationConfig(
                    max_output_tokens=2000,
                    temperature=0.3
                )
            )
            return response.text.strip()
        except Exception as e:
            logging.error(f"Gemini API call failed: {str(e)}")
            raise LLMServiceError(f"Gemini API call failed: {str(e)}", {"model": model_name})
    
    def _parse_json_response(self, response: str, eval_type: str) -> Dict[str, Any]:
        """Parse JSON response from LLM"""
        try:
            # Try to extract JSON from response
            start_idx = response.find('{')
            end_idx = response.rfind('}') + 1
            if start_idx != -1 and end_idx != 0:
                json_str = response[start_idx:end_idx]
                return json.loads(json_str)
            else:
                raise ValueError("No JSON found in response")
        except Exception as e:
            print(f"Failed to parse LLM response: {str(e)}")
            # Return mock data as fallback
            if eval_type == "cv":
                return self._mock_cv_evaluation()
            else:
                return self._mock_project_evaluation()
    
    def _mock_cv_evaluation(self) -> Dict[str, Any]:
        """Mock CV evaluation for testing"""
        return {
            "technical_skills_match": 0.82,
            "experience_level": 0.75,
            "relevant_achievements": 0.68,
            "cultural_fit": 0.85,
            "cv_match_rate": 0.82,
            "cv_feedback": "Strong in backend and cloud, limited AI integration experience."
        }
    
    def _mock_project_evaluation(self) -> Dict[str, Any]:
        """Mock project evaluation for testing"""
        return {
            "correctness": 4.2,
            "code_quality": 3.8,
            "resilience": 3.5,
            "documentation": 4.0,
            "creativity_bonus": 3.2,
            "project_score": 3.7,
            "project_feedback": "Meets prompt chaining requirements, lacks error handling robustness."
        }

# Global instance
llm_service = LLMService()