requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
google-generativeai==0.3.2
chromadb==0.4.18
python-docx==1.1.0
PyPDF2==3.0.1
aiofiles==23.2.0
asyncio==3.4.3
uuid==1.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4